Change sensor read output to hexadecimal numbers
Changes 'am2320.h' -- updates the function definitions' parameter types.

Changes 'am2320.c' -- updates the functions' logic to return and process
integer values, instead of floats.

Changes 'main.c' -- updates the sensor-data-processing functions to
handle the new data type.

This commit changes the numbering system of the communication frame
values to hexadecimal.

+++ Sem III/Microprocessor-programming/AM2320-sensor/Core/Inc/am2320_it.h
@@ -38,7 +38,7 @@ extern TIM_HandleTypeDef htim6;
 void AM2320_Init(AM2320_HandleTypeDef *am2320, I2C_HandleTypeDef *hi2c, uint8_t sensor_address);
 void AM2320_InitSensorRead(AM2320_HandleTypeDef *am2320);
 void AM2320_ReadSensorData(AM2320_HandleTypeDef *am2320);
-void AM2320_ProcessSensorData(AM2320_HandleTypeDef *am2320, float *temperature, float *humidity);
+void AM2320_ProcessSensorData(AM2320_HandleTypeDef *am2320, uint16_t *temperature, uint16_t *humidity);
 void AM2320_ReadSensorInfo(AM2320_HandleTypeDef *am2320);
 void AM2320_ProcessSensorInfo(AM2320_HandleTypeDef *am2320, uint32_t *am2320_id, uint16_t *am2320_model, uint8_t *am2320_version);
 uint16_t compute_CRC(uint8_t *frame_data, uint16_t data_length);
